+++ src/components/core/tab-group/tab-group.tsx
@@ -1,3 +1,5 @@
+import "./tab-group.scss";
+
 import { Tab } from "@headlessui/react";
 import cx from "classnames";
 import { ReactNode, useRef, useState } from "react";
@@ -6,8 +8,6 @@ import { usePatientHistory } from "@/components/content/patient-history/use-pati
 import { withErrorBoundary } from "@/components/core/error-boundary";
 import { ListBox } from "@/components/core/list-box/list-box";
 import { useBreakpoints } from "@/hooks/use-breakpoints";
-import "./tab-group.scss";
-import { PatientRefreshHistoryMessageStatus } from "@/services/patient-history/patient-history-types";
 
 export type TabGroupProps = {
   children?: ReactNode;
@@ -64,8 +64,7 @@ function TabGroupComponent({
       className={cx(className, "ctw-tab-group ctw-scrollable-pass-through-height ctw-w-full")}
     >
       <PatientHistoryStatus
-        messages={patientHistoryDetails.latestServiceMessages}
-        status={patientHistoryDetails.lastStatus as PatientRefreshHistoryMessageStatus}
+        status={patientHistoryDetails.lastStatus}
         date={patientHistoryDetails.lastRetrievedAt}
       />
       <Tab.Group selectedIndex={selectedTabIndex} onChange={handleOnChange}>